move models and constants to their own file
