+++ apps/report-execution/src/libraries/support/pa_01/calculations.py
@@ -1,0 +1,3 @@
+class CaseAssignmentsAndOutcomes:
+    def build():
+        pass
